Move team assignments code into a component
Created a component called 'MashDashboard' to move the team assignments code to so it is easier to test. Alongside this I made adjustments to the 'MashCard' code so it includes an Action button.
